add, abstraction for image processor and theme manger and apply refactors
