+++ app/src/main/java/com/sanaa/tudee_assistant/MainActivity.kt
@@ -4,12 +4,9 @@ import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
-import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
-import com.sanaa.tudee_assistant.data.utils.ThemeManager
+import com.sanaa.tudee_assistant.domain.ThemeManager
 import com.sanaa.tudee_assistant.presentation.TudeeApp
 import com.sanaa.tudee_assistant.presentation.design_system.theme.TudeeTheme
 import org.koin.compose.koinInject
@@ -20,13 +17,7 @@ class MainActivity : ComponentActivity() {
         enableEdgeToEdge()
         setContent {
             val themeManager: ThemeManager= koinInject()
-
-            var isDark by remember { mutableStateOf(false) }
-            LaunchedEffect(themeManager) {
-                themeManager.isDarkTheme.collect {
-                    isDark = it
-                }
-            }
+            val isDark by themeManager.isDarkTheme.collectAsState(initial = false)
 
             TudeeTheme (isDark){
                 TudeeApp()
