+++ app/src/main/java/com/sanaa/tudee_assistant/data/utils/ImageProcessorImpl.kt
@@ -7,16 +7,17 @@ import android.net.Uri
 import android.os.Build
 import android.provider.MediaStore
 import androidx.core.graphics.scale
+import com.sanaa.tudee_assistant.domain.ImageProcessor
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
 import java.io.File
 import java.io.FileOutputStream
 import java.io.IOException
 
-class ImageProcessor(private val context: Context) {
+class ImageProcessorImpl(private val context: Context) : ImageProcessor{
 
     @Throws(IOException::class)
-    suspend fun processImage(uri: Uri): Bitmap = withContext(Dispatchers.IO) {
+    override suspend fun processImage(uri: Uri): Bitmap = withContext(Dispatchers.IO) {
         try {
             val originalBitmap = if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.P) {
                 ImageDecoder.decodeBitmap(ImageDecoder.createSource(context.contentResolver, uri))
@@ -32,7 +33,7 @@ class ImageProcessor(private val context: Context) {
     }
 
     @Throws(IOException::class)
-    suspend fun saveImageToInternalStorage(bitmap: Bitmap): String = withContext(Dispatchers.IO) {
+    override suspend fun saveImageToInternalStorage(bitmap: Bitmap): String = withContext(Dispatchers.IO) {
         val directory = File(context.filesDir, STORAGE_CATEGORIES_FILE_DIR).apply {
             if (!exists()) mkdirs()
         }
