+++ app/src/main/java/com/sanaa/tudee_assistant/data/utils/ThemeManagerImpl.kt
@@ -6,24 +6,25 @@ import androidx.datastore.preferences.core.Preferences
 import androidx.datastore.preferences.core.booleanPreferencesKey
 import androidx.datastore.preferences.core.edit
 import androidx.datastore.preferences.preferencesDataStore
+import com.sanaa.tudee_assistant.domain.ThemeManager
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
 
 
-class ThemeManager(private val context: Context) {
+class ThemeManagerImpl(private val context: Context) : ThemeManager{
 
      companion object {
         private val Context.dataStore: DataStore<Preferences> by preferencesDataStore("storeData")
         val DARK_THEME_KEY = booleanPreferencesKey("is_dark_theme")
     }
 
-    val isDarkTheme: Flow<Boolean> = context.dataStore.data
+    override val isDarkTheme: Flow<Boolean> = context.dataStore.data
         .map { prefrences->
             prefrences[DARK_THEME_KEY]?:false
         }
 
 
-    suspend fun setDarkTheme(enabled: Boolean) {
+    override suspend fun setDarkTheme(enabled: Boolean) {
         context.dataStore.edit { prefs ->
             prefs[DARK_THEME_KEY] = enabled
         }
