+++ app/src/main/java/com/sanaa/tudee_assistant/di/imageProcessorModule.kt
@@ -1,8 +1,9 @@
 package com.sanaa.tudee_assistant.di
 
-import com.sanaa.tudee_assistant.data.utils.ImageProcessor
+import com.sanaa.tudee_assistant.data.utils.ImageProcessorImpl
+import com.sanaa.tudee_assistant.domain.ImageProcessor
 import org.koin.dsl.module
 
 val imageProcessorModule = module {
-    single { ImageProcessor(get()) }
+    single < ImageProcessor>{ ImageProcessorImpl(get()) }
 }
