+++ app/src/main/java/com/sanaa/tudee_assistant/di/themeManagerModule.kt
@@ -1,8 +1,9 @@
 package com.sanaa.tudee_assistant.di
 
-import com.sanaa.tudee_assistant.data.utils.ThemeManager
+import com.sanaa.tudee_assistant.data.utils.ThemeManagerImpl
+import com.sanaa.tudee_assistant.domain.ThemeManager
 import org.koin.dsl.module
 
 val themeManagerModule = module {
-        single { ThemeManager(get()) }
+        single<ThemeManager> { ThemeManagerImpl(get()) }
 }
