+++ app/src/main/java/com/sanaa/tudee_assistant/domain/ImageProcessor.kt
@@ -1,0 +1,9 @@
+package com.sanaa.tudee_assistant.domain
+
+import android.graphics.Bitmap
+import android.net.Uri
+
+interface ImageProcessor {
+    suspend fun processImage(uri: Uri): Bitmap
+    suspend fun saveImageToInternalStorage(bitmap: Bitmap): String
+}
