+++ app/src/main/java/com/sanaa/tudee_assistant/domain/ThemeManager.kt
@@ -1,0 +1,8 @@
+package com.sanaa.tudee_assistant.domain
+
+import kotlinx.coroutines.flow.Flow
+
+interface ThemeManager {
+    val isDarkTheme: Flow<Boolean>
+    suspend fun setDarkTheme(enabled: Boolean)
+}
