+++ app/src/main/java/com/sanaa/tudee_assistant/presentation/screen/main/MainScreen.kt
@@ -4,7 +4,6 @@ import androidx.annotation.DrawableRes
 import androidx.compose.foundation.layout.Column
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Modifier
 import androidx.navigation.NavDestination.Companion.hasRoute
 import androidx.navigation.NavGraph.Companion.findStartDestination
@@ -14,13 +13,8 @@ import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.currentBackStackEntryAsState
 import androidx.navigation.compose.rememberNavController
-import com.sanaa.tudee_assistant.data.utils.ThemeManager
 import com.sanaa.tudee_assistant.presentation.design_system.component.TudeeBottomNavBar
 import com.sanaa.tudee_assistant.presentation.design_system.component.TudeeBottomNavBarItem
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.launch
-import org.koin.compose.getKoin
-import org.koin.compose.koinInject
 
 internal var navItems = mutableListOf<NavItem>()
