+++ app/src/main/java/com/sanaa/tudee_assistant/presentation/utils/CategoryThumbnail.kt
@@ -1,12 +1,8 @@
-package com.sanaa.tudee_assistant.data.utils
+package com.sanaa.tudee_assistant.presentation.utils
 
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
 import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.unit.dp
 import coil.compose.AsyncImage
 import com.sanaa.tudee_assistant.R
